requirements.txt
rdkit-pypi
ipython
joblib
numpy
pandas
py-spy
pytest
scikit-learn==1.1.2
scipy
tornado
tqdm
matplotlib==3.5.2
seaborn==0.11.2
xmltodict